requirements.txt
Django==1.9.4
argparse==1.2.1
dj-database-url==0.4.0
dj-static==0.0.6
django-bootstrap3==7.0.1
django-formtools==1.0
django-registration-redux==1.4
django-toolbelt==0.0.1
gunicorn==19.4.5
psycopg2==2.6.1
static3==0.6.1
wsgiref==0.1.2
